services/retriever/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
faiss-cpu>=1.8.0
spacy>=3.7.0
numpy>=1.26.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
prometheus-client>=0.20.0
httpx>=0.27.0
